made size float too.
git-svn-id: https://btanks.svn.sourceforge.net/svnroot/btanks/trunk@402 5c0bf991-d514-0410-a65f-ce4832e571f3

+++ src/base_object.h
@@ -28,7 +28,7 @@ namespace sdlx {
 
 class BaseObject : public mrt::Serializable {
 public:
-	v3<int> size;
+	v3<float> size;
 	float mass, speed, ttl, impassability;
 	int hp;
